Add game cleaning command. Add printing of messages.

+++ PolyEngine/Scripts/common/project_mgr/project_base.py
@@ -32,9 +32,17 @@ class ProjectBase():
         self._save()
 
     def clean_build(self):
-        shutil.rmtree(self._build_path)
-        shutil.rmtree(self._dist_path)
-        shutil.rmtree(self._common_build_path)
+        print(f"Cleaning build files in {self._project_name}")
+        if os.path.exists(self._build_path):
+            print(f"Removing Build directory in {self._project_name}")
+            shutil.rmtree(self._build_path)
+        # CommonBuild is not present in games
+        if os.path.exists(self._common_build_path):
+            print(f"Removing CommonBuild directory in {self._project_name}")
+            shutil.rmtree(self._common_build_path)
+        if os.path.exists(self._dist_path):
+            print(f"Removing Dist directory in {self._project_name}")
+            shutil.rmtree(self._dist_path)
 
     @property
     def name(self):
